fix hooks and add npm-run-all to all packages;
