+++ example/public/index.html
@@ -12,7 +12,9 @@
     <div id="content">
       <h1 id="page-title">WalletConnect</h1>
 
-      <div id="page-info"></div>
+      <div id="page-info">
+        <a href="/web3-provider.html">Switch to web3-provider example</a>
+      </div>
 
       <div id="page-actions">
         <p>Click to Connect</p>
